+++ BLL/Services/Impl/OrderService.cs
@@ -4,7 +4,7 @@ namespace BLL.Services.Impl
 {
     public class OrderService
     {
-        public int Order(ISubscriptionService subscription, int value)
+        public float Order(ISubscriptionService subscription, int value)
         {
             return subscription.Pay(value);
         }
